Move getConnection to connection
