+++ src/custom/components/Header/NetworkSelector/NetworkSelectorMod.tsx
@@ -1,6 +1,6 @@
 import { Trans } from '@lingui/macro'
 import { useWeb3React } from '@web3-react/core'
-import { getConnection } from '@cow/modules/wallet/web3-react/types'
+import { getConnection } from '@cow/modules/wallet/web3-react/connection'
 import { getChainInfo } from 'constants/chainInfo'
 import { CHAIN_IDS_TO_NAMES, SupportedChainId } from 'constants/chains'
 import useParsedQueryString from 'hooks/useParsedQueryString'
